iseq() for arrays:  removed template parameter
replaced template parameter with
iterator_traits<Arra*>::difference_type.  i am not sure yet whether this
is portable, because that maybe typedef can be for T*.  I'm not sure if
this is allowed by the standard, but if it is then this change can still
introduce ambiguities if i have another overload that takes two
parameters of two template parameters (as i plan to have, for two
iterators).  this change is because otherwise i for sure have
ambiguity issues.